requirements.txt
easyocr
opencv-python
pdf2image
Pillow
numpy
tk
